rework repeatables and groups - groups now working yet
