+++ interfaces/field.ts
@@ -1,23 +1,16 @@
 export default interface Field {
+  add_more_label: string;
+  add_more_position: 'bottom' | 'top';
   attributes?: {
     rows?: number | null;
   };
   checked_value?: string;
+  children?: Array<Field>;
   field_class?: string;
   first_empty?: boolean;
-  fm?: {
-    add_more_label: string;
-    add_more_position: 'bottom' | 'top';
-    attributes: {
-      rows: number | null;
-    };
-    children: Array<Field>;
-    field_class: string;
-    label: string;
-    limit: number,
-    name: string;
-  };
   label?: string;
+  limit: number,
+  minimumCount?: number,
   name: string;
   options?: string[],
   title?: string;
