+++ slotfills/sections/group.tsx
@@ -1,31 +1,5 @@
-import React from 'react';
-import { Button, PanelRow } from '@wordpress/components';
-import { SortableContainer, SortableElement, SortableHandle } from 'react-sortable-hoc'; // TODO: replace with https://github.com/clauderic/dnd-kit
-import { arrayMoveImmutable } from 'array-move';
-import { v4 as uuidv4 } from 'uuid';
-import { __ } from '@wordpress/i18n';
-
 import FMObject from '@/interfaces/fm-object';
 import FieldRouter from './fieldRouter';
-import AddMoreButton from '../../components/add-more-button';
-
-import './group.scss';
-
-interface ChildProps {
-  children?: React.ReactNode | React.ReactNode[];
-}
-
-const SortableItem = SortableElement(({ children }: ChildProps) => (
-  <li>
-    {children}
-  </li>
-));
-
-const SortableList = SortableContainer(({ children }: ChildProps) => (
-  <ul className="fm-gutenberg-sortable-list">{children}</ul>
-));
-
-const DragHandle = SortableHandle(() => <span className="fm-gutenberg-move-handle" aria-label={__('Move', 'fm-gutenberg')}>::</span>);
 
 interface GroupProps {
   field: {
@@ -42,10 +16,6 @@ interface GroupProps {
 
 export default function Group({
   field: {
-    add_more_label: addMoreLabel = '',
-    add_more_position: addMorePosition = 'bottom',
-    limit = null,
-    minimumCount = null,
     name,
     children = [],
   },
@@ -63,83 +33,19 @@ export default function Group({
     return [indexValue, setIndexValue];
   };
 
-  const addNew = () => {
-    const newValueArray = [...value, {}];
-    setValue(newValueArray);
-  };
-
-  const removeElement = (index: number) => {
-    const newValueArray = [...value];
-    newValueArray.splice(index, 1);
-    setValue(newValueArray);
-  };
-
-  const onSortEnd = ({ oldIndex, newIndex }: { oldIndex: number, newIndex: number }) => {
-    const newValue = arrayMoveImmutable([...value], oldIndex, newIndex);
-    setValue(newValue);
-  };
-
   return (
-    <>
-      {addMorePosition === 'top' ? (
-        <AddMoreButton
-          addMoreLabel={addMoreLabel}
-          addNew={addNew}
-          limit={limit}
-          minimumCount={minimumCount}
-        />
-      ) : null}
-      <PanelRow>
-        <SortableList
-          onSortEnd={onSortEnd}
-          useDragHandle
-        >
-          {value ? value.map((childValue: number | string | FMObject, index: number) => {
-            const key = uuidv4();
-            return (
-              <SortableItem
-                key={key}
-                index={index}
-              >
-                <PanelRow>
-                  <div className="fm-gutenberg-panel-container">
-                    <div className="fm-gutenberg-controls">
-                      <DragHandle />
-                      <Button
-                        onClick={() => removeElement(index)}
-                        className="fm-gutenberg-remove"
-                      >
-                        <span className="screen-reader-text">{__('Remove', 'fm-gutenberg')}</span>
-                      </Button>
-                    </div>
-                    <div>
-                      {Object.keys(children).map((itemKey: keyof typeof children) => {
-                        const child = children[itemKey];
-                        return (
-                          <FieldRouter
-                            field={child}
-                            index={index}
-                            valueHook={useIndexedValue}
-                            key={itemKey}
-                          />
-                        );
-                      })}
-                    </div>
-                  </div>
-                </PanelRow>
-              </SortableItem>
-            );
-          }) : null}
-        </SortableList>
-      </PanelRow>
-      {addMorePosition === 'bottom' ? (
-        <AddMoreButton
-          addMoreLabel={addMoreLabel}
-          addNew={addNew}
-          limit={limit}
-          minimumCount={minimumCount}
-        />
-      ) : null}
-    </>
+    <div>
+      {Object.keys(children).map((itemKey: keyof typeof children) => {
+        const child = children[itemKey];
+        return (
+          <FieldRouter
+            field={child}
+            // index={index}
+            valueHook={useIndexedValue}
+            key={itemKey}
+          />
+        );
+      })}
+    </div>
   );
 }
